backend/requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
# sqlalchemy==2.0.23
# alembic==1.13.0
# psycopg2-binary==2.9.9
# redis==5.0.1
# celery==5.3.4
motor==3.3.2
pymongo[srv]
beanie==1.23.6
slither-analyzer==0.11.3

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# AI Integration
openai==1.3.5
tiktoken==0.5.1

# File Processing
python-magic==0.4.27
aiofiles==23.2.1

# Utilities
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
email-validator==2.2.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

markdown==3.5.2  
jinja2==3.1.2